oh fuck

+++ cpp/fake_rand.h
@@ -2,4 +2,4 @@
 
 void FakeRandInit();
 
-int FakeRand();
+size_t FakeRand();
